Chore: unit tests
