+++ src/lib/__tests__/PageControls-test.html
@@ -1,0 +1,1 @@
+<div id="test-page-controls-container"> </div>
