+++ src/lib/viewers/doc/PresentationViewer.js
@@ -2,7 +2,6 @@ import autobind from 'autobind-decorator';
 import throttle from 'lodash.throttle';
 import DocBaseViewer from './DocBaseViewer';
 import PresentationPreloader from './PresentationPreloader';
-import PageControls from '../../PageControls';
 import { CLASS_INVISIBLE } from '../../constants';
 import {
     ICON_FILE_PRESENTATION,
@@ -191,8 +190,8 @@ class PresentationViewer extends DocBaseViewer {
         this.controls.add(__('zoom_out'), this.zoomOut, 'bp-exit-zoom-out-icon', ICON_ZOOM_OUT);
         this.controls.add(__('zoom_in'), this.zoomIn, 'bp-enter-zoom-in-icon', ICON_ZOOM_IN);
 
-        this.pageControls = new PageControls(this.controls, this.prevPage, this.nextPage);
         this.pageControls.init(this.pdfViewer.pagesCount);
+        this.pageControls.addListener('setpage', this.setPage);
 
         this.controls.add(
             __('enter_fullscreen'),
